NP-1106: move exceptions outside database package
