+++ src/main/java/no/unit/nva/database/DatabaseServiceImpl.java
@@ -12,9 +12,9 @@ import com.amazonaws.services.dynamodbv2.model.ComparisonOperator;
 import com.amazonaws.services.dynamodbv2.model.Condition;
 import java.util.List;
 import java.util.Optional;
-import no.unit.nva.database.exceptions.InvalidInputRoleException;
-import no.unit.nva.database.exceptions.InvalidRoleInternalException;
-import no.unit.nva.database.exceptions.InvalidUserInternalException;
+import no.unit.nva.exceptions.InvalidInputRoleException;
+import no.unit.nva.exceptions.InvalidRoleInternalException;
+import no.unit.nva.exceptions.InvalidUserInternalException;
 import no.unit.nva.database.intefaces.WithType;
 import no.unit.nva.model.RoleDto;
 import no.unit.nva.model.UserDto;
