+++ src/main/java/no/unit/nva/exceptions/InvalidUserInternalException.java
@@ -1,4 +1,4 @@
-package no.unit.nva.database.exceptions;
+package no.unit.nva.exceptions;
 
 import nva.commons.exceptions.ApiGatewayException;
 import org.apache.http.HttpStatus;
