+++ src/main/java/no/unit/nva/handlers/AddRoleHandler.java
@@ -4,10 +4,10 @@ import com.amazonaws.services.lambda.runtime.Context;
 import java.util.Optional;
 import no.unit.nva.database.DatabaseService;
 import no.unit.nva.database.DatabaseServiceImpl;
-import no.unit.nva.database.exceptions.DataHandlingError;
-import no.unit.nva.database.exceptions.InvalidInputRoleException;
-import no.unit.nva.database.exceptions.InvalidRoleInternalException;
-import no.unit.nva.database.exceptions.UnexpectedException;
+import no.unit.nva.exceptions.DataHandlingError;
+import no.unit.nva.exceptions.InvalidInputRoleException;
+import no.unit.nva.exceptions.InvalidRoleInternalException;
+import no.unit.nva.exceptions.UnexpectedException;
 import no.unit.nva.model.RoleDto;
 import nva.commons.handlers.ApiGatewayHandler;
 import nva.commons.handlers.RequestInfo;
