+++ src/test/java/no/unit/nva/model/RoleDtoTest.java
@@ -12,7 +12,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import no.unit.nva.database.exceptions.InvalidRoleInternalException;
+import no.unit.nva.exceptions.InvalidRoleInternalException;
 import no.unit.nva.model.RoleDto.Builder;
 import nva.commons.utils.attempt.Failure;
 import org.junit.jupiter.api.Test;
